extraction_from_bbox/requirements.txt
pymongo==4.6.3
psycopg2-binary==2.9.9
shapely==2.0.4
pykml==0.2.0
